refactor: reorganize utoo-web source structure
