+++ packages/utoo-web/src/index.ts
@@ -2,5 +2,5 @@ export {
   compatOptionsFromWebpack,
   type WebpackConfig,
 } from "@utoo/pack-shared";
-export { Project } from "./project";
-export * from "./type";
+export { Project } from "./project/Project";
+export * from "./types";
