+++ packages/utoo-web/src/utils/installServiceWorker.ts
