+++ packages/utoo-web/src/webpackLoaders/loaderWorkerPool.ts
@@ -1,7 +1,7 @@
 import * as comlink from "comlink";
-import { internalEndpoint } from "../internalProject";
-import * as sabcom from "../sabcom";
-import { Binding } from "../type";
+import { internalEndpoint } from "../project/InternalProject";
+import * as sabcom from "../utils/sabcom";
+import { Binding } from "../types";
 import initWasm, {
   Project as ProjectInternal,
   registerWorkerScheduler,
@@ -9,7 +9,7 @@ import initWasm, {
   WebWorkerTermination,
   workerCreated,
 } from "../utoo";
-import { LoaderRunnerMeta } from "./type";
+import { LoaderRunnerMeta } from "./types";
 
 let nextWorkerId = 0;
